Fix: refactor classes name for Button and Chip components
