+++ ouds_core/lib/components/button/internal/ouds_button_border_modifier.dart
@@ -11,12 +11,12 @@
 //
 
 import 'package:flutter/material.dart';
-import 'package:ouds_core/components/button/internal/button_loading_modifier.dart';
+import 'package:ouds_core/components/button/internal/ouds_button_loading_modifier.dart';
 import 'package:ouds_core/components/button/ouds_button.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
 
 /// Used to apply a border with color, width and radius associated to the hierarchy
-class ButtonBorderModifier {
+class OudsButtonBorderModifier {
   static WidgetStateProperty<BorderSide?> resolveBorderColor(
     BuildContext context,
     OudsButtonHierarchy hierarchy,
@@ -25,7 +25,7 @@ class ButtonBorderModifier {
     return WidgetStateProperty.resolveWith<BorderSide?>(
       (Set<WidgetState> states) {
         if (style == OudsButtonStyle.loading) {
-          return ButtonLoadingModifier.getBorderColor(context, hierarchy);
+          return OudsButtonLoadingModifier.getBorderColor(context, hierarchy);
         }
         if (states.contains(WidgetState.pressed)) {
           return _getPressedBorderColor(context, hierarchy);
