+++ ouds_core/lib/components/chip/internal/ouds_chip_text_style_modifier.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/cupertino.dart';
-import 'package:ouds_core/components/chip/internal/ouds_chip_control_state.dart';
+import 'package:ouds_core/components/chip/internal/ouds_ouds_chip_control_state.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
 
 class OudsChipControlTextColorModifier {
